General-purpose function to print padded columns
The so-called "keys" will be right-padded with spaces so that they align
to the longest key string, and printed in the left-hand column.

The so-called "values" will be left-padded with spaces so that they
align to the longest value string, and printed in the right-hand column.

The two columns will be separated by " - ", but this can be easily
changed.